Use int64 in sparse index dtype test to ensure actual mismatch

+++ tests/test_helpers.py
@@ -253,7 +253,7 @@ def test_assert_equal_sparse_index_dtype(attr):
     """assert_equal(exact=True) should detect indptr/indices dtype mismatches."""
     a = sparse.csr_matrix(np.eye(3))
     b = sparse.csr_matrix(np.eye(3))
-    setattr(b, attr, getattr(b, attr).astype(np.int32))
+    setattr(b, attr, getattr(b, attr).astype(np.int64))
 
     # Non-exact comparison should pass (values are identical)
     assert_equal(a, b, exact=False)
